==> Aim 3_Dobbs/code/02_pull in BRFSS data for aim 3.Rmd <==
---
title: "02_pull in BRFSS data for aim 3"
author: "Andrea Molino"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: html_document
---

This Rmarkdown pulls in the necessary BRFSS data for Andrea Molino's dissertation Aim 3 to create the analytical dataset with applicable variables and such. This website was used to guide my data download process: https://michaelminn.net/tutorials/r-brfss/ 

```{r, include=FALSE, echo=FALSE, message=FALSE, warning=FALSE}
rm(list=ls())
knitr::opts_chunk$set(echo = TRUE)

### Run state categorizations Rmarkdown to create the state categorizations needed for the BRFSS data
rmarkdown::render("01_aim 3 state categorizations.Rmd")
rm(data_states_min, state_categories_map, state_categories_map_bottom, states_shape, states_shape_join)

### Pull in additional libraries 
library(crosstable)
library(dplyr)
library(tidyverse)
library(ggplot2)
library(data.table)
library(table1)
library(readr)
library(readxl)

### BRFSS data downloaded from the following webpages as ASCII files: 
# 2022: https://www.cdc.gov/brfss/annual_data/annual_2022.html
# 2020: https://www.cdc.gov/brfss/annual_data/annual_2020.html
# 2018: https://www.cdc.gov/brfss/annual_data/annual_2018.html
# 2016: https://www.cdc.gov/brfss/annual_data/annual_2016.html


### Via same links above, "Variable Layout" pages were copied into CSVs, and will be read into R as well so that correct variables are lined up with raw ASCII files 
columns_2022 = read.csv("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/variable layouts/brfss_var_layout_2022.csv")
columns_2020 = read.csv("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/variable layouts/brfss_var_layout_2020.csv")
columns_2018 = read.csv("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/variable layouts/brfss_var_layout_2018.csv")
columns_2016 = read.csv("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/variable layouts/brfss_var_layout_2016.csv")


### Function to add create File_Width column to each data frame, which indicates the length of a variable is in the raw ASCII file 
process_columns <- function(data_frame) {
  data_frame$File_Width <- sapply(1:nrow(data_frame), function(y) {
    if (y < nrow(data_frame)) {
      data_frame$X...Starting_Column[y + 1] - data_frame$X...Starting_Column[y]
    } else {
      1
    }
  })
  
  data_frame <- data_frame[data_frame$File_Width > 0, ]
  return(data_frame)
}

# List of variable layout data frames to process 
columns_all <- list(columns_2022, columns_2020, columns_2018, columns_2016) 

# Run variable layout data frames (columns) through the process_columns function, and assign the processed data frames back to their original names to be able to identify which BRFSS year the variables are associated with
processed_columns <- lapply(columns_all, process_columns)

columns_2022 <- processed_columns[[1]]
columns_2020 <- processed_columns[[2]]
columns_2018 <- processed_columns[[3]]
columns_2016 <- processed_columns[[4]]


### Pull in BRFSS data and align with file widths established above
# Define the column positions based on the starting and ending points, and then read the fixed-width files with the column positions
column_positions_2022 <- fwf_widths(columns_2022$File_Width, col_names = columns_2022$Variable_Name)
column_positions_2020 <- fwf_widths(columns_2020$File_Width, col_names = columns_2020$Variable_Name)
column_positions_2018 <- fwf_widths(columns_2018$File_Width, col_names = columns_2018$Variable_Name)
column_positions_2016 <- fwf_widths(columns_2016$File_Width, col_names = columns_2016$Variable_Name)

brfss_2022 <- read_fwf("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/LLCP2022ASC/LLCP2022.ASC", col_positions = column_positions_2022)
brfss_2020 <- read_fwf("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/LLCP2020ASC/LLCP2020.ASC", col_positions = column_positions_2020)
brfss_2018 <- read_fwf("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/LLCP2018ASC/LLCP2018.ASC", col_positions = column_positions_2018)
brfss_2016 <- read_fwf("C:/Users/amoli/OneDrive - UW/Dissertation/Aim 3_Dobbs/BRFSS data/LLCP2016ASC/LLCP2016.ASC", col_positions = column_positions_2016)

orig_brfss_2022 <- brfss_2022
orig_brfss_2020 <- brfss_2020
orig_brfss_2018 <- brfss_2018
orig_brfss_2016 <- brfss_2016


### Drop itmes in environment that are no longer useful, aka everything besides our actual BRFSS data 
rm(columns_2022, columns_2020, columns_2018, columns_2016, columns_all, 
   processed_columns, column_positions_2022, column_positions_2020, column_positions_2018, column_positions_2016)


```
