==> Aim 3_Dobbs/code/04_DID checks.Rmd <==
---
title: "04_DID checks"
author: "Andrea Molino"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    df_print: paged
    theme: cosmo
    toc: yes
---

This Rmarkdown was created to run the "difference-in-differences" checks needed for my third dissertation aim, and to also provide some of the information for my SER 2025 abstract. 

Rmarkdown files 01, 02, and 03 create the necessary `brfss_final` dataset used in this Rmarkdown, which is the final cleaned BRFSS data. As of December 2024, it solely includes 2018, 2020, and 2022 data, but will eventually include 2016 and 2024.  


```{r, include=FALSE, message=FALSE, warning=FALSE}
### Set up 
rm(list=ls())
knitr::opts_chunk$set(echo = TRUE)

library(table1)
library(crosstable)
library(dplyr)
library(tidyverse)
library(ggplot2)
library(data.table)
library(tibble)
library(purrr)
library(labelled)
library(grid)
library(gridExtra)
library(lubridate)
library(plm)
library(broom)

load(file="../data/brfss_final.RData")


```


```{r, include=FALSE, message=FALSE, warning=FALSE}
### Additional CCS compliant variable with "missing" explicitly coded 
brfss_final <- brfss_final %>% 
  mutate(ccs_cmplnt_miss = case_when(
    is.na(ccs_cmplnt) ~ "CCS information missing", 
    TRUE ~ ccs_cmplnt
  ))


### Create 0/1 ccs_cmplnt variable
brfss_final <- brfss_final %>%
  mutate(ccs_cmplnt_y1n0 = case_when(
    ccs_cmplnt == "Compliant" ~ 1, 
    ccs_cmplnt == "Noncompliant" ~ 0, 
    TRUE ~ NA
  ))


### Create a dataset where everyone is non-missing for CCS compliant status AND all of the data are pre-Dobbs ruling (aka pre June 24, 2022)
brfss_final_ccsdata_predobbs <- subset(brfss_final, !is.na(ccs_cmplnt) & prejune22_y1n0==1)

```


# Basic descriptive statistics 

Below, I am providing some descriptive statistics for the entire study population. I look at both the "entire" study population (anyone within the inclusion criteria in 2018, 2020, or 2022) as well as restricted to those who have non-missing cervical cancer screening information, meaning about 15% are excluded. 

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### Basic descriptive statistics 
table1(~ factor(pre0post1dobbs) + 
         factor(year_brfss) + 
         age_cat + 
         age_cat_collap +
         edu_full + 
         edu_collap + 
         urban1rural2 + 
         factor(prejune22_y1n0) + 
         postdobbs_cat +
         ccs_cmplnt, 
       data = brfss_final, caption = "Study population characteristics: entire population")

table1(~ factor(pre0post1dobbs) + 
         factor(year_brfss) + 
         age_cat + 
         age_cat_collap +
         edu_full + 
         edu_collap + 
         urban1rural2 + 
         factor(prejune22_y1n0) + 
         postdobbs_cat +
         ccs_cmplnt, 
       data = brfss_final_ccsdata_predobbs, caption = "Study population characteristics: those with non-missing CCS information and pre-June 24, 2022")

table1(~ factor(pre0post1dobbs) + 
         factor(year_brfss) + 
         age_cat + 
         age_cat_collap +
         edu_full + 
         edu_collap + 
         urban1rural2 + 
         factor(prejune22_y1n0) +
         postdobbs_cat |
         ccs_cmplnt_miss, 
       data = brfss_final, caption = "Study population characteristics: stratified by CCS status")

table1(~ factor(pre0post1dobbs) + 
         ccs_cmplnt_miss + 
         age_cat + 
         age_cat_collap +
         edu_full + 
         edu_collap + 
         urban1rural2 + 
         factor(prejune22_y1n0) +
         postdobbs_cat |
         factor(year_brfss), 
       data = brfss_final, caption = "Study population characteristics: stratified by BRFSS year")

table1(~ factor(pre0post1dobbs) + 
         ccs_cmplnt_miss + 
         age_cat + 
         age_cat_collap +
         edu_full + 
         edu_collap + 
         urban1rural2 + 
         factor(prejune22_y1n0) + 
         postdobbs_cat |
         factor(year_brfss), 
       data = brfss_final_ccsdata_predobbs, caption = "Study population characteristics: stratified by BRFSS year")


```


\pagebreak

Here I am looking at the number of participants in each state, and in each type of state (abortion banned vs. abortion accessible post-Dobbs).

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### State statistics 
table1(~ state_name, 
       data = brfss_final, caption = "Distribution of states in analysis overall")

table1(~ postdobbs_cat | 
         factor(year_brfss), 
       data = brfss_final, caption = "Distribution of post-Dobbs category by BRFSS survey year")

table1(~ state_name | 
         factor(year_brfss), 
       data = brfss_final, caption = "Distribution of state by BRFSS survey year")


### State statistics: non-missing CCS status
table1(~ state_name, 
       data = brfss_final_ccsdata_predobbs, caption = "Distribution of states in analysis overall: only those with non-missing CCS status and pre-June 24, 2022")

table1(~ postdobbs_cat | 
         factor(year_brfss), 
       data = brfss_final_ccsdata_predobbs, caption = "Distribution of post-Dobbs category by BRFSS survey year: only those with non-missing CCS status and pre-June 24, 2022")

table1(~ state_name | 
         factor(year_brfss), 
       data = brfss_final_ccsdata_predobbs, caption = "Distribution of state by BRFSS survey year: only those with non-missing CCS status and pre-June 24, 2022")


### Number of abortion restricted vs. accessible states 
state_cats <- brfss_final_ccsdata_predobbs %>% 
  distinct(state_name, postdobbs_cat) 

table1(~ state_name | postdobbs_cat, 
       data = state_cats, caption = "Abortion accessible vs. banned status")


```

\pagebreak

# Difference in differences checks 

Here, I am starting to run my difference-in-differences checks for my third dissertation aim. These cheks are 1) test of prior trends, 2) placebo test, and 3) set of balancing tests for changes in state-level composition. All of these primarily come guidance provided by Wing et al. 2018 "Designing Difference in Difference Studies: Best Practices for Public Health Policy Research" (link found here: https://pubmed.ncbi.nlm.nih.gov/29328877/). 

For these checks, I am ONLY including those who have non-missing information on cervical cancer screening whose BRFSS interview happened in the pre-Dobbs time period (aka, pre-June 24, 2022). I may eventually change this, but for now, this is what I am doing for validity. 


## Test of prior trends

I am first just plotting the change in cervical cancer screening compliance proportion over time comparing abortion banned vs. accessible states.  

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### Assessment of change in outcome over time comparing abortion banned vs. accessible states, as well as overall change over time 
prop_ccs_cmplnt_year <- 
  brfss_final_ccsdata_predobbs %>%
  group_by(postdobbs_cat, year_brfss) %>% 
  summarise(mean_ccs_cmplnt = mean(ccs_cmplnt_y1n0))

prop_ccs_cmplnt_year_overall <- 
  brfss_final_ccsdata_predobbs %>%
  group_by(year_brfss) %>% 
  summarise(mean_ccs_cmplnt = mean(ccs_cmplnt_y1n0))

fig_ccs_year_dobbs <- 
  ggplot(prop_ccs_cmplnt_year, aes(x=year_brfss, y=(mean_ccs_cmplnt)*100)) + 
  geom_point(aes(col=postdobbs_cat)) +
  geom_line(aes(col=postdobbs_cat)) +
  theme_bw() + 
  scale_x_continuous(limits=c(2017.5, 2022.5), breaks=c(2018,2020,2022), labels=c("2018","2020","2022")) +
  scale_y_continuous(limits=c(0,100), breaks=c(0,25,50,75,100), labels=c("0%","25%","50%","75%","100%")) +
  scale_color_manual(values=c("#009E73", "#D55E00"), name = "Post-Dobbs abortion access status") +
  theme(plot.title=element_text(hjust=0.5), legend.position="bottom") + 
  labs(title="Cervical cancer screening compliance\nby post-Dobbs abortion access status", 
       y="% cervical cancer screening compliant", x="BRFSS survey year",
       caption="Only participants who completed their BRFSS\nsurvey prior to June 24, 2022 included")

print(fig_ccs_year_dobbs)
print(prop_ccs_cmplnt_year)
print(prop_ccs_cmplnt_year_overall)

ggsave(file="../figures/cervical cancer screening over BRFSS wave for DID.pdf", plot=fig_ccs_year_dobbs, width=6, height=4)
ggsave(file="../figures/cervical cancer screening over BRFSS wave for DID.png", plot=fig_ccs_year_dobbs, width=6, height=4)
ggsave(file="../figures/cervical cancer screening over BRFSS wave for DID.jpg", plot=fig_ccs_year_dobbs, width=6, height=4)

```

Below is the test of prior trends. I first run a DID model with the interaction term (aka the DID estimate), and then fit the same model but without the DID estimate interaction term. This allows me to run an ANOVA and get an F-test to quantify if both interaction terms are significantly different than 0 (since my `year_brfss` variable is 2018 vs. 2020 vs 2022)

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### Test of prior trends
model_priortrends_interaction <- 
  glm(ccs_cmplnt_y1n0 ~ factor(postdobbs_cat) + factor(year_brfss) + factor(postdobbs_cat)*factor(year_brfss), 
      data = brfss_final_ccsdata_predobbs, 
      family = binomial(link = "logit"))

model_priortrends_nointeraction <- 
  glm(ccs_cmplnt_y1n0 ~ factor(postdobbs_cat) + factor(year_brfss),
      data = brfss_final_ccsdata_predobbs, 
      family = binomial(link = "logit"))

summary(model_priortrends_interaction)
summary(model_priortrends_nointeraction)

anova(model_priortrends_nointeraction, model_priortrends_interaction)

```


## Placebo test 

Here, I am running two placebo tests: one where I 'pretend' that Dobbs was implemented in 2019, so 2018 is considered pre-intervention and 2020/2022 are post intervention, and another where I 'pretend' that Dobbs was implemented in 2021, so 2018/2020 are pre-intervention and 2022 is post-intervention. This essentialy allows me to see if the something else in time could have contributed to a difference in the outcome prior to the intervention being implemented. 

Below is the "pretend the Dobbs decision happened in 2019" test. 

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### Pretend the Dobbs decision happened in 2019 and run DID model 
brfss_final_ccsdata_predobbs <- brfss_final_ccsdata_predobbs %>% 
  mutate(placebo_2019_pre0post1 = case_when(
    year_brfss %in% c(2018)       ~ 0, 
    year_brfss %in% c(2020, 2022) ~ 1, 
    TRUE ~ 99
  ))

model_placebo_2019 <- 
  glm(ccs_cmplnt_y1n0 ~ factor(postdobbs_cat) + factor(placebo_2019_pre0post1) + factor(postdobbs_cat)*factor(placebo_2019_pre0post1), 
      data = brfss_final_ccsdata_predobbs, 
      family = binomial(link = "logit"))

model_placebo_2019_results <- tidy(model_placebo_2019, exponentiate = TRUE, conf.int = TRUE)

summary(model_placebo_2019)
print(model_placebo_2019_results)


```

And now below is the "pretend the Dobbs decision happened in 2021" test. 

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### Pretend the Dobbs decision happened in 2021 and run DID model 
brfss_final_ccsdata_predobbs <- brfss_final_ccsdata_predobbs %>% 
  mutate(placebo_2021_pre0post1 = case_when(
    year_brfss %in% c(2018, 2020) ~ 0, 
    year_brfss %in% c(2022)       ~ 1, 
    TRUE ~ 99
  ))

model_placebo_2021 <- 
  glm(ccs_cmplnt_y1n0 ~ factor(postdobbs_cat) + factor(placebo_2021_pre0post1) + factor(postdobbs_cat)*factor(placebo_2021_pre0post1), 
      data = brfss_final_ccsdata_predobbs, 
      family = binomial(link = "logit"))

model_placebo_2021_results <- tidy(model_placebo_2021, exponentiate = TRUE, conf.int = TRUE)

summary(model_placebo_2021)
print(model_placebo_2021_results)


```


## Balancing tests for changes in state-level composition

Here, I am running a set of balancing tests for changes in state/group-level composition. Essentially, the goal is to investigate if there were any age, education level, or urban/rural county changes from 2018 vs. 2020 vs. 2022 comparing states that did and did not ban abortion. 

While this test is typically supposed to be performed with pre- and post-intervention data, to see if changes in composition could be the reason for the DID difference, I am looking at pre-Dobbs changes over time to see if there is any evidence at first. 

The first test here is the age group test, where the outcome is a binary age variable (30-39 years vs. 40-49 years).

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### Balancing test: AGE GROUP 
brfss_final_ccsdata_predobbs <- brfss_final_ccsdata_predobbs %>% 
  mutate(age_cat_collap_young0old1 = case_when(
    age_cat_collap %in% c("30-39") ~ 0, 
    age_cat_collap %in% c("40-49") ~ 1, 
    is.na(age_cat_collap)          ~ NA,
    TRUE ~ 99
  ))

table(brfss_final_ccsdata_predobbs$age_cat_collap, brfss_final_ccsdata_predobbs$age_cat_collap_young0old1, useNA="always")

model_balancing_age_interaction <- 
  glm(age_cat_collap_young0old1 ~ factor(postdobbs_cat) + factor(year_brfss) + factor(postdobbs_cat)*factor(year_brfss), 
      data = brfss_final_ccsdata_predobbs, 
      family = binomial(link = "logit"))

model_balancing_age_nointeraction <- 
  glm(age_cat_collap_young0old1 ~ factor(postdobbs_cat) + factor(year_brfss), 
      data = brfss_final_ccsdata_predobbs,
      family = binomial(link = "logit"))

summary(model_balancing_age_interaction)
summary(model_balancing_age_nointeraction)

anova(model_balancing_age_nointeraction, model_balancing_age_interaction)

```

The second test is the education level test, where the outcome is a binary education variable (did not graduate from college or technical school vs. graduated from college or technical school).

```{r, echo=FALSE, message=FALSE, warning=FALSE}
### Balancing test: EDUCATION LEVEL
brfss_final_ccsdata_predobbs <- brfss_final_ccsdata_predobbs %>% 
  mutate(edu_collap_less0more1 = case_when(
    edu_collap %in% c("Did not graduate from college or technical school") ~ 0, 
    edu_collap %in% c("Graduated from college or technical school")        ~ 1, 
    is.na(edu_collap)                                                      ~ NA,
    TRUE ~ 99
  ))

table(brfss_final_ccsdata_predobbs$edu_collap, brfss_final_ccsdata_predobbs$edu_collap_less0more1, useNA="always")

model_balancing_edu_interaction <- 
  glm(edu_collap_less0more1 ~ factor(postdobbs_cat) + factor(year_brfss) + factor(postdobbs_cat)*factor(year_brfss), 
      data = brfss_final_ccsdata_predobbs, 
      family = binomial(link = "logit"))

model_balancing_edu_nointeraction <- 
  glm(edu_collap_less0more1 ~ factor(postdobbs_cat) + factor(year_brfss), 
      data = brfss_final_ccsdata_predobbs,
      family = binomial(link = "logit"))

summary(model_balancing_edu_interaction)
summary(model_balancing_edu_nointeraction)

anova(model_balancing_edu_nointeraction, model_balancing_edu_interaction)

```

The third test is the urban/rural county level test, where the outcome is a binary living in an urban vs. rural county variable. 

```{r, echo=FALSE, message=FALSE, warning=FALSE}

### Balancing test: URBAN/RURAL STATUS
brfss_final_ccsdata_predobbs <- brfss_final_ccsdata_predobbs %>% 
  mutate(urban0rural1 = case_when(
    urban1rural2 %in% c("Urban county") ~ 0, 
    urban1rural2 %in% c("Rural county") ~ 1, 
    is.na(urban1rural2)                 ~ NA,
    TRUE ~ 99
  ))

table(brfss_final_ccsdata_predobbs$urban1rural2, brfss_final_ccsdata_predobbs$urban0rural1, useNA="always")

model_balancing_urb_interaction <- 
  glm(urban0rural1 ~ factor(postdobbs_cat) + factor(year_brfss) + factor(postdobbs_cat)*factor(year_brfss), 
      data = brfss_final_ccsdata_predobbs,
      family = binomial(link = "logit"))

model_balancing_urb_nointeraction <- 
  glm(urban0rural1 ~ factor(postdobbs_cat) + factor(year_brfss), 
      data = brfss_final_ccsdata_predobbs,
      family = binomial(link = "logit"))

summary(model_balancing_urb_interaction)
summary(model_balancing_urb_nointeraction)

anova(model_balancing_urb_nointeraction, model_balancing_urb_interaction)

```



```{r, include=FALSE, message=FALSE, warning=FALSE}


```
